+++ app/src/main/java/com/example/yabamiru/MainActivity.kt
@@ -9,7 +9,7 @@ import android.widget.ArrayAdapter
 import android.widget.Toast
 import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity(), RecyclerViewHolder.ItemClickListener  {
+class MainActivity : AppCompatActivity(), RecyclerAdapter.RecyclerViewHolder.ItemClickListener  {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
