+++ app/src/main/res/layout/activity_tag.xml
@@ -9,7 +9,10 @@
         android:layout_margin="8dp"
         card_view:cardElevation="2dp"
         android:foreground="?android:attr/selectableItemBackground"
-        app:cardBackgroundColor="@color/colorCardTag">
+        app:cardBackgroundColor="@color/colorCardTag"
+        app:contentPaddingLeft="4dp"
+        app:contentPaddingRight="4dp">
+
     <TextView
             android:id="@+id/tag_text"
             android:layout_width="wrap_content"
