+++ app/src/main/res/layout/main_list_row.xml
@@ -13,9 +13,7 @@
 
     <android.support.constraint.ConstraintLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="horizontal"
-            android:gravity="center_vertical">
+            android:layout_height="match_parent">
 
         <ImageView
                 android:id="@+id/main_row_imageView"
